pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trend-opportunity-bot"
version = "0.1.0"
description = "Collect trend signals and generate opportunity hypothesis cards"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1",
  "pydantic>=2.6.0",
  "typer>=0.12.0",
  "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.6.0",
]

[project.scripts]
trendbot = "trendbot.cli:app"

[tool.setuptools.packages.find]
include = ["trendbot*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
